curriculo-lattes-backend/requirements.txt
Flask==2.3.0
flask_api==3.1
flask_cors==4.0.0
mysql-connector-python
xmltodict==0.13.0
Werkzeug==2.3.0
networkx==3.3
matplotlib==3.9.0
PyQt6